將model中功能flag的欄位資料型態改為long

+++ GodPay-CMS/Services/DTO/UserFilterRsp.cs
@@ -30,7 +30,7 @@ namespace GodPay_CMS.Services.DTO
         /// <summary>
         /// 權限旗標
         /// </summary>
-        public int Func { get; set; }
+        public long Func { get; set; }
 
         /// <summary>
         /// 目前狀態
